requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
joblib>=1.3.0
openpyxl>=3.1.0
groq>=0.4.0
plotly>=5.18.0
python-dotenv>=1.0.0